GLDraw: new features
git-svn-id: svn://ultimatepp.org/upp/trunk@12729 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/GLDraw/Triangles.cpp
@@ -10,7 +10,7 @@ void GLTriangles::Draw(const GLContext2D& dd)
 	static GLCode program(R"(
 	#version 130
 	    attribute vec3 p;
-	    attribute vec3 c;
+	    attribute uvec3 c;
 		uniform vec2 offset;
 		uniform vec2 scale;
 		varying vec4 v_color;
